Update weather_app.py
In this updated version, we've made the following changes:

Added an extra parameter temp_unit to the get_weather_data() function, which allows the user to specify the preferred temperature unit (Celsius or Fahrenheit). The default is set to Celsius ("C").
In the main() function, we prompt the user to enter their preferred temperature unit ("C" or "F"). If an invalid unit is provided, it defaults to Celsius.
The script will now ask the user to enter their preferred temperature unit. Based on the user's input, it will display the temperature in the chosen unit in the weather data output.